requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
Django==4.2.20
django-fernet-fields-v2==0.9
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
importlib_metadata==8.7.0
Markdown==3.8
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
phonenumbers==9.0.4
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
zipp==3.21.0
